% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR_auxiliary_functionsI.R
\name{collate_clusters}
\alias{collate_clusters}
\title{Collate sequence IDs from an existing clustering according to a new,
given clustering of the existing clusters}
\usage{
collate_clusters(to_clust, orig_clust)
}
\arguments{
\item{to_clust}{A list giving clustering of factors. In other words this is
the clustering of clusters of sequences given in `orig_clust`}

\item{orig_clust}{A list of sequence IDs in existing clusters}
}
\value{
A list with sequence IDs collated by the specified clustering
}
\description{
Collate sequences original divided across n clusters into a
 new set of m clusters. These `m` clusters obtained by clustering the
 original `n` clusters.
Assume a collection of 100 sequences across seven existing
clusters. These seven clusters are collated to obtain three resulting
clusters. Collating 100 sequences distributed across the seven clusters into
the resulting three clusters can be achieved with collate clusters.
}
\examples{



set.seed(123)
n <- 7; nn <- 100
orig_clust_labels <- ceiling(n * runif(nn))
orig_clust <- seqArchR::get_seqs_clust_list(orig_clust_labels)

to_clust <- list(c(1,4), c(2,3,5), c(6,7))

collate_clusters(to_clust = to_clust, orig_clust = orig_clust)

}
