% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqArchR_auxiliary_functionsI.R
\name{get_clBasVec}
\alias{get_clBasVec}
\alias{get_clBasVec_k}
\alias{get_clBasVec_m}
\alias{get_seqClLab}
\title{Get functions for seqArchR result object}
\usage{
get_clBasVec(res, iter)

get_clBasVec_k(res, iter)

get_clBasVec_m(res, iter)

get_seqClLab(res, iter = NULL)
}
\arguments{
\item{res}{seqArchR result object.}

\item{iter}{Choose the iteration of seqArchR result to get from.}
}
\value{
get_clBasVec A list with two elements `nBasisVectors` (integer) and
`basisVectors` (matrix).

get_clBasVec_k The number of basis vectors (integer).

get_clBasVec_m The basis vectors' matrix with features along the
rows of the matrix.

get_seqClLab A character vector denoting the cluster IDs for
each sequence.
}
\description{
Basis vectors' information for the selected iteration.
}
\section{Functions}{
\itemize{
\item \code{get_clBasVec_k}: Get the number of basis vectors (clusters)
at the  selected iteration.

\item \code{get_clBasVec_m}: The basis vectors matrix at the selected
iteration. Note that eatures along rows.

\item \code{get_seqClLab}: Get the cluster IDs for each sequence. Note that
order of sequences here is as per the input.
}}

\examples{
res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

k <- get_clBasVec_k(res=res, iter=2)

bMat <- get_clBasVec_m(res=res, iter=2)

## cluster labels of sequences from final clustering
scLab <- get_seqClLab(res=res, iter=2)

}
\seealso{
\code{\link{seqs_str}}
}
