% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_FASTA.R
\name{get_one_hot_encoded_seqs}
\alias{get_one_hot_encoded_seqs}
\title{Get one-hot encoded sequences}
\usage{
get_one_hot_encoded_seqs(seqs, sinuc_or_dinuc = "sinuc")
}
\arguments{
\item{seqs}{A \code{\link[Biostrings]{DNAStringSet}} object holding the
given DNA sequences}

\item{sinuc_or_dinuc}{character string, 'sinuc' or 'dinuc' to select for
mono- or dinucleotide profiles.}
}
\value{
A sparse matrix of sequences represented with one-hot-encoding
}
\description{
Get the one-hot encoding representation of the given sequences.
}
\examples{

fname <- system.file("extdata", "example_data.fa.gz",
                        package = "seqArchR", mustWork = TRUE)


rawSeqs <- prepare_data_from_FASTA(fasta_fname = fname,
                        raw_seq = TRUE)

seqs_dinuc <- get_one_hot_encoded_seqs(seqs = rawSeqs,
                                       sinuc_or_dinuc = "dinuc")

}
\seealso{
\code{\link{prepare_data_from_FASTA}} for generating one-hot
encoding of sequences from a FASTA file

Other input functions: 
\code{\link{prepare_data_from_FASTA}()}
}
\concept{input functions}
