% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsI.R
\name{get_seqs_clust_list}
\alias{get_seqs_clust_list}
\title{Retrieve sequence clusters as a list from the sequence labels}
\usage{
get_seqs_clust_list(seqs_clust_lab)
}
\arguments{
\item{seqs_clust_lab}{Sequences with cluster labels as in the seqArchR
result object.}
}
\value{
A list holding sequence IDs belonging in each cluster.
}
\description{
Given the sequence cluster labels from the seqArchR result
object, returns the clusters separated as a list.
}
\examples{

clustLabels <- sample(seq_len(4), 50, replace = TRUE)
print(clustLabels)
get_seqs_clust_list(clustLabels)

}
