% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archR_auxiliary_functionsII.R
\name{handle_dir_creation}
\alias{handle_dir_creation}
\title{Handle directory creation}
\usage{
handle_dir_creation(o_dir, vrbs)
}
\arguments{
\item{o_dir}{Specify the output directory name with its complete path.}

\item{vrbs}{Set verbosity to TRUE or FALSE}
}
\value{
The (updated) dir name
}
\description{
Given the output directory name with its complete path, this
function checks if a directory of same name exists the given location. If
yes, then it adds a suffix (a number) to the given directory name, and
proceeds to create the directory.
}
