% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_ggheatmap}
\alias{plot_ggheatmap}
\title{Plot a given position weight matrix as a heatmap using ggplot2}
\usage{
plot_ggheatmap(pwm_mat, pos_lab = NULL, pdf_name = NULL, fixed_coord = FALSE)
}
\arguments{
\item{pwm_mat}{Matrix (usually a PWM, but can be non-normalized/any matrix)
to be represented as a heatmap.}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the
positions are labeled from 1 to the length of the sequences.}

\item{pdf_name}{Name of the file which will be saved as PDF.}

\item{fixed_coord}{Set this to TRUE to use a fixed aspect ratio for the
plot. Default is FALSE.}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later. If \code{pdf_name} is given, it is also saved and the
\code{ggplot} object returned.
}
\description{
The given matrix is plotted as a heatmap using \code{ggplot2}'s
\code{geom_tile}.
}
\examples{

res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

pwm <- seqArchR::make_dinuc_PWMs(get_clBasVec_m(res,iter=1)[,1],
                        add_pseudo_counts = FALSE)

plot_ggheatmap(pwm_mat = pwm, fixed_coord = TRUE)

}
\seealso{
\code{\link{plot_ggseqlogo}} for plotting PWMs as sequence logos

Other visualization functions: 
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_bas_vec_heatmap_seqlogo}()},
\code{\link{viz_seqs_acgt_mat}()}
}
\concept{visualization functions}
