% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_ggseqlogo}
\alias{plot_ggseqlogo}
\title{Visualize a given (PWM) matrix as a sequence logo.}
\usage{
plot_ggseqlogo(
  pwm_mat,
  method = "bits",
  pos_lab = NULL,
  pdf_name = NULL,
  bits_yax = "full",
  fixed_coord = FALSE
)
}
\arguments{
\item{pwm_mat}{Matrix (usually a PWM, but can be any non-normalized matrix)
to be represented as a sequence logo. Rownames must be letters.}

\item{method}{For \code{ggseqlogo}; either of 'custom', 'bits', or
'probability'. Default is 'bits'.}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the
positions are labeled from 1 to the length of the sequences.}

\item{pdf_name}{Name of the file which will be saved as PDF.}

\item{bits_yax}{Specify 'full' if the information content y-axis limits
should be 0-2 or 'auto' for a suitable limit. The 'auto' setting adjusts
the y-axis limits according to the maximum information content of the
sequence logo. Default is 'full'.}

\item{fixed_coord}{Set this to TRUE to use a fixed aspect ratio for the
plot. Default is FALSE.}
}
\value{
A ggplot2 object so you can simply call \code{print} or \code{save}
on it later. If \code{pdf_name} is given, it is also saved in addition to
returning the ggplot object.
}
\description{
Visualize a given (PWM) matrix as a sequence logo.
}
\examples{

res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

pwm <- seqArchR::make_dinuc_PWMs(get_clBasVec_m(res,iter=1)[,1],
                        add_pseudo_counts = FALSE)

plot_ggseqlogo(pwm_mat = pwm, fixed_coord = TRUE)

}
\seealso{
\code{\link{plot_ggheatmap}} for plotting PWMs as heatmaps,
\code{\link{plot_ggseqlogo_of_seqs}} for visualizing a collection of
sequences by their sequence logo.

Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{viz_bas_vec_heatmap_seqlogo}()},
\code{\link{viz_seqs_acgt_mat}()}
}
\concept{visualization functions}
