% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggheatmap.R
\name{plot_pwm}
\alias{plot_pwm}
\title{Visualize a position weight matrix as a heatmap or sequence logo}
\usage{
plot_pwm(
  pwm_mat,
  method = "heatmap",
  pos_lab = NULL,
  pdf_name = NULL,
  fixed_coord = FALSE,
  bits_yax = "full"
)
}
\arguments{
\item{pwm_mat}{Matrix (usually a PWM, but can be any non-normalized matrix)
to be visualized. Rownames must be letters.}

\item{method}{Character. Set this to 'heatmap' when plotting a heatmap, else
you can set it to either of 'custom', 'bits', or 'probability'. Default is
'heatmap'.}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the
positions are labeled from 1 to the length of the sequences.}

\item{pdf_name}{Name of the file which will be saved as PDF.}

\item{fixed_coord}{Set this to TRUE to use a fixed aspect ratio for the
plot. Default is FALSE.}

\item{bits_yax}{Specify 'full' if the information content y-axis limits
should be 0-2 or 'auto' for a suitable limit. The 'auto' setting adjusts
the y-axis limits according to the maximum information content of the
sequence logo. Default is 'full'.}
}
\value{
A ggplot object so you can simply call \code{print} or \code{save}
on it later. If \code{pdf_name} is given, it is also saved and the
\code{ggplot2} object returned.
}
\description{
The given position weight matrix is plotted as a heatmap
or sequence logo
}
\examples{

res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

pwm <- seqArchR::make_PWMs(get_clBasVec_m(res,iter=1)[,1],
                        add_pseudo_counts = FALSE, sinuc = FALSE)

plot_pwm(pwm_mat = pwm, method = "heatmap", fixed_coord = TRUE)

plot_pwm(pwm_mat = pwm, method = "bits", fixed_coord = TRUE)

}
\seealso{
\code{\link{plot_ggseqlogo_of_seqs}} for visualizing a collection of
sequences by their sequence logo.

Other visualization functions: 
\code{\link{viz_bas_vec}()},
\code{\link{viz_seqs_acgt_mat}()}
}
\concept{visualization functions}
