% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_basis_vectors.R
\name{viz_bas_vec_heatmap_seqlogo}
\alias{viz_bas_vec_heatmap_seqlogo}
\alias{viz_bas_vec_seqlogo}
\alias{viz_bas_vec_heatmap}
\title{Visualize the NMF basis vectors in a paired heatmap and sequence logo
plot}
\usage{
viz_bas_vec_heatmap_seqlogo(
  feat_mat,
  method = "bits",
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc",
  fixed_coord = FALSE
)

viz_bas_vec_seqlogo(
  feat_mat,
  method = "bits",
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc",
  fixed_coord = FALSE
)

viz_bas_vec_heatmap(
  feat_mat,
  pos_lab = NULL,
  add_pseudo_counts = FALSE,
  pdf_name = NULL,
  sinuc_or_dinuc = "sinuc",
  fixed_coord = FALSE
)
}
\arguments{
\item{feat_mat}{The features matrix (basis vectors matrix) from seqArchR.}

\item{method}{For \code{ggseqlogo} -- either of "custom", "bits", or
"probability". Default is "bits".}

\item{pos_lab}{Labels for sequence positions, should be of same
length as that of the sequences. Default value is NULL, when the
positions are labeled from 1 to the length of the sequences.}

\item{add_pseudo_counts}{Logical, taking values TRUE or FALSE, default
set to FALSE. Setting it to TRUE will enable adding pseudo-counts to the
features matrix.}

\item{pdf_name}{Filename to save the plot, also provide the extension.}

\item{sinuc_or_dinuc}{"sinuc" or "dinuc" for choosing between mono- and
dinucleotide profiles respectively.}

\item{fixed_coord}{Set this to TRUE to use a fixed aspect ratio for the
plot. Default is FALSE.}
}
\value{
nothing
}
\description{
The given features matrix is visualized as a heatmap followed
by a sequence logo where the positions are aligned for better
visualization.
}
\section{Functions}{
\itemize{
\item \code{viz_bas_vec_seqlogo}: Visualize the NMF basis vectors
as a sequence logo

\item \code{viz_bas_vec_heatmap}: Visualize the
NMF basis vectors as a heatmap
}}

\examples{
res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

viz_bas_vec_heatmap_seqlogo(feat_mat = get_clBasVec_m(res,iter=1),
                            sinuc_or_dinuc = "dinuc", fixed_coord = TRUE)

viz_bas_vec_seqlogo(feat_mat = get_clBasVec_m(res,iter=1),
                     sinuc_or_dinuc = "dinuc", fixed_coord = TRUE)

# Visualizing basis vector for a single cluster
viz_bas_vec_heatmap(feat_mat = as.matrix(get_clBasVec_m(res,iter=1)[,3]),
                     sinuc_or_dinuc = "dinuc", fixed_coord = TRUE)

}
\seealso{
Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_seqs_acgt_mat}()}
}
\concept{visualization functions}
