% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_matrix_of_acgt_image.R
\name{viz_seqs_acgt_mat}
\alias{viz_seqs_acgt_mat}
\title{Visualize raw DNA sequences as an image}
\usage{
viz_seqs_acgt_mat(
  seqs,
  pos_lab = NULL,
  xt_freq = min(length(pos_lab), 5),
  yt_freq = min(length(seqs), 100),
  use_col = c("darkgreen", "blue", "orange", "red"),
  add_legend = TRUE,
  use_legend = Biostrings::DNA_BASES,
  save_fname = NULL,
  file_type = "PNG",
  f_width = 450,
  f_height = 900,
  f_units = "px"
)
}
\arguments{
\item{seqs}{The sequences as a DNAStringSet object.}

\item{pos_lab}{The labels to be used for the sequence positions.
Default: Sequence positions are labeled from 1 to the length of the
sequences.}

\item{xt_freq}{The x-axis tick frequency. Expects a positive integer less
than the length of the sequences. Default is 5.}

\item{yt_freq}{The y-axis tick frequency. Expects a positive integer less
than number of sequences. Default is 100.}

\item{use_col}{A vector of four colors used for the DNA bases A, C, G,
and T (in that order).}

\item{add_legend}{Logical. Whether legend should be added to the plot.
Default is TRUE.}

\item{use_legend}{A character vector of letters in the input sequences.
Default is \code{\link[Biostrings]{DNA_BASES}}, used for DNA sequences.}

\item{save_fname}{Specify the filename (with extension) for saving the
plot to disk.}

\item{file_type}{Specify the file type, namely PNG, JPEG, TIFF.}

\item{f_width}{Specify the width for the plot. This depends on the length of
sequences.}

\item{f_height}{Specify the height for the plot. This depends on the number
of sequences.}

\item{f_units}{Specify the units in which the height and width are given.}
}
\value{
Nothing returned to the R interpreter.
}
\description{
This function plots the collection of sequences as an image
matrix.
}
\examples{

res <- readRDS(system.file("extdata", "example_seqArchRresult.rds",
         package = "seqArchR", mustWork = TRUE))

# Image matrix of sequences in the input order
viz_seqs_acgt_mat(seqs = seqs_str(res))

# Image matrix of sequences ordered by the clustering from seqArchR
use_seqs <- seqs_str(res, iter = NULL, cl = NULL, ord = TRUE)
viz_seqs_acgt_mat(seqs = use_seqs)

# Image matrix of sequences belonging to a single cluster
use_seqs <- seqs_str(res, iter = 2, cl = 2)
viz_seqs_acgt_mat(seqs = use_seqs)

}
\seealso{
Other visualization functions: 
\code{\link{plot_ggheatmap}()},
\code{\link{plot_ggseqlogo}()},
\code{\link{viz_bas_vec_heatmap_seqlogo}()}
}
\concept{visualization functions}
